better clustering
- Show geohashes on client side in dev mode
- Make pixi circles semi-transparent for better visibility when stacked
- Auto mode toggle
- Update button
- Disable popup when in edit mode
- add geohash to rust backend
- add better fc support to bootstrapping
- Add a whole new clustering mode that's much better than the "fast mode"
- Remove the old trash

+++ server/src/models/api.rs
@@ -52,7 +52,7 @@ pub enum ReturnType {
 pub struct RouteGeneration {
     pub instance: Option<String>,
     pub radius: Option<f64>,
-    pub min_points: Option<i32>,
+    pub min_points: Option<usize>,
     pub generations: Option<usize>,
     pub devices: Option<usize>,
     pub data_points: Option<Vec<GenericData>>,
